#69v2 updated program api ( name instead of id)
