+++ course_admin/serializers.py
@@ -33,3 +33,19 @@ class CourseSerializer(serializers.ModelSerializer):
         instance.updated_by = self.context['request'].user    
         instance.save()
         return instance
+
+
+class ProgramSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Program
+        fields = '__all__'
+        read_only_fields = ('created_by', 'updated_by')
+
+    def create(self, validated_data):
+        validated_data['created_by'] = self.context['request'].user
+        validated_data['updated_by'] = self.context['request'].user
+        return super().create(validated_data)
+
+    def update(self, instance, validated_data):
+        validated_data['updated_by'] = self.context['request'].user
+        return super().update(instance, validated_data)
