+++ course_admin/urls.py
@@ -5,4 +5,6 @@ urlpatterns = [
     path('courses/', CourseCreateView.as_view(), name='course-create'),     
     path('courses/<str:course_name>/', CourseCreateView.as_view(), name='course-detail'),  
     path('courses/update/<str:course_name>/', CourseUpdateView.as_view(), name='course-update'),
+    path('programs/', ProgramCreateAPIView.as_view(), name='program-list-create'),
+    path('programs/<str:name>/', ProgramRetrieveUpdateDestroyAPIView.as_view(), name='program-retrieve-update-destroy'),
 ]
