+++ courses/serializers.py
@@ -2,7 +2,6 @@ from rest_framework import serializers
 from .models import (
     Course,
     Batch,
-    Program
 )
 
 class RetrieveProgramsSerializer(serializers.Serializer):
@@ -49,13 +48,3 @@ class BatchSerializer(serializers.ModelSerializer):
         
 
 
-class ProgramSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Program
-        fields = '__all__'  
-        read_only_fields = ('created_by', 'updated_by')  
-
-    def create(self, validated_data):
-        validated_data['created_by'] = self.context['request'].user
-        validated_data['updated_by'] = self.context['request'].user
-        return super().create(validated_data)
