+++ courses/views.py
@@ -2,7 +2,7 @@ from .serializers import(
     RetrieveProgramsSerializer,
     RetrieveCourseSerializer,
     BatchSerializer,
-    ProgramSerializer
+    
 )
 from .models import(
     Program,
@@ -95,17 +95,3 @@ class RetrieveCourseBatchesView(ListAPIView):
         return Batch.objects.filter(start_date__gt=current_date,course__name__iexact=course_name)
 
     
-class ProgramCreateAPIView(generics.CreateAPIView):
-    queryset = Program.objects.all()
-    serializer_class = ProgramSerializer
-    permission_classes = [IsCourseAdmin]
-
-    def perform_create(self, serializer):
-        serializer.save(
-            created_by=self.request.user,
-            updated_by=self.request.user
-        )
-
-class ProgramRetrieveUpdateDestroyAPIView(generics.RetrieveUpdateDestroyAPIView):
-    queryset = Program.objects.all()
-    serializer_class = ProgramSerializer
